ORM/requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.22
numpy==1.19.4
pandas==1.1.4
plotly==4.12.0
protobuf==3.13.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.20
